Update me-DevOpsDays.html
I deleted some bullet points and edited a few others. I wasn't sure what you meant with the unicorns so I took them out. If you want to keep them in, maybe clarify what you mean?  Also -- what are you trying to say with this point:  "The audience really appreciated the opportunity to talk about feedback. The room full of mostly guys needed and appreciated the chance to listen, talk and think about this topic. Soft skills might be the hardest to master." This comes off as a bit of a slight towards the conference attendees, to me. Are you saying that they were bad at listening and having a discussion with each other? Or were *they* saying that the teams they work with are lacking in the soft skills.

Good job, and you clearly learned a TON from that conference.